+++ src/operators/skipWhile/skipWhile.js
@@ -1,0 +1,11 @@
+exports.skipWhile = predicate => function*(source) {
+    let hasFailed = false;
+    for (const value of source) {
+        hasFailed = hasFailed || !predicate(value);
+        if (hasFailed) {
+            yield value;
+        } else {
+            continue;
+        }
+    }
+}