+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - changed the drop operation to skip
 - changed the take operation to limit
 - added a function to produce a reusable iterable from a factory function
+- added a skipWhile operator